+++ app/src/main/java/at/ac/tuwien/caa/docscan/logic/Helper.java
@@ -1,5 +1,6 @@
 package at.ac.tuwien.caa.docscan.logic;
 
+import android.app.Activity;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
@@ -27,6 +28,7 @@ import java.io.File;
 import java.io.FileFilter;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
+import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Comparator;
@@ -615,21 +617,24 @@ public class Helper {
 //        tempFileName = tempFileName.substring(0, dotPos) + "-temp" + tempFileName.substring(dotPos);
 
         File originalFile = new File(fileName);
+        if (originalFile == null)
+            return false;
+
         File tempFile;
         try {
             tempFile = File.createTempFile("img", ".jpg", originalFile.getParentFile());
         } catch (IOException e) {
             e.printStackTrace();
             Crashlytics.log("Helper.saveMat");
+            Log.d(CLASS_NAME, "Helper.saveMat");
             Crashlytics.logException(e);
             return false;
         }
 
+        if (tempFile == null)
+            return false;
+
         String tempFileName = tempFile.getAbsolutePath();
-
-        Log.d(CLASS_NAME, "tempFileName: " + tempFileName);
-
-//        File tempFile = new File(tempFileName);
 
         boolean isSaved = false;
 
@@ -648,7 +653,7 @@ public class Helper {
             Crashlytics.logException(e);
         } finally {
 //            Delete the temporary file, if it still exists:
-            if (tempFile.exists())
+            if (tempFile != null && tempFile.exists())
                 tempFile.delete();
         }
